+++ packages/joy-media/src/DiscoveryProvider.tsx
@@ -8,7 +8,7 @@ import { Vec } from '@polkadot/types';
 import { Url } from '@joystream/types/discovery'
 import ApiContext from '@polkadot/react-api/ApiContext';
 import { ApiProps } from '@polkadot/react-api/types';
-import { JoyInfo } from '@polkadot/joy-utils/JoyWarn';
+import { JoyInfo } from '@polkadot/joy-utils/JoyStatus';
 
 export type BootstrapNodes = {
   bootstrapNodes?: Url[],
